added schema validation
