+++ src/pages/contacts/components/ContactInput/styles.module.css
@@ -58,6 +58,15 @@
     display: none;
 }
 
+.error_contacts {
+    margin-top: 8px;
+    min-height: 16px;
+    font-size: 1em;
+    line-height: 14px;
+    font-weight: 400;
+    color: var(--colors-interface-orange);
+  }
+
 @media (max-width: 600px) {
     .container {
         padding-bottom: 0;
