+++ src/pages/contacts/components/contacts-form/styles.module.css
@@ -1,0 +1,13 @@
+.form {
+  padding: 0 30px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  row-gap: 20px;
+}
+
+@media (max-width: 600px) {
+  .form {
+    padding: 0;
+  }
+}
